Complete two KIF tests, refine test convention in the process.
- Add accessibility aspect to related views, include
localization. Add additional boilerplate as needed. The
convention for view controllers is to dedicate a method for
this `-setAccessibilityLabels` that usually runs on
`-viewDidLoad`.

- Add additional methods on the test actor for checking view
text, as long as it has the proper getter. This requires
storing the reference text in the test method itself. This
also requires a helper `-viewWithAccessibilityLabel:` based
on base test actor code.

Other:

- Put test actor additions on KIFUITestActor instead.
- Global test imports.
- Refactor out defines for accessibility labels. Save them
as properties on the test case itself. They should readable.
- Rename `CoreFlowTests` -> `CoreFunctionalTests`.