Rework Add-Event aspect in Months screen...
Should be triggered by tapping any empty space
(i.e. backgroundView). No more 'cloaked cells'
workarounds.

Also provide visual feedback when background gets
tapped. The caveat is by the time we confirm the
tap gesture there's already a feedback delay. To
avoid the delay we assume a tap. So the feedback
is really any interaction on top of empty space.
We need to use UIScrollViewDelegate methods to
ensure compatibility, since the view controller
isn't and can't be the delegate for the pan
recognizer.

Also note the aspect comment tags.