Rework ETTransitionManager...
Refactor and rename ETTransitionManager into a
more specialized zoom transition manager and stop
using the class as a singleton / with shared
state. Get rid of some extra properties. Update
affected view controllers.

Note we haven't renamed the class to preserve the
diff. Also note the zoom prefixes for our own
properties are to prevent conflicts with the
delegate methods. Lastly, note for the navigation
controller we can just return the
transitioningDelegate if possible for the required
delegate methods.