Months screen should update after event saves...
The initial requirement for this was just a save
notification aspect and proper invalidation and
updating of existing event data. Instead of
re-fetching all events, the saved event is just
added to the existing collection, which then gets
resorted. Any custom collections get invalidated.

And this is where the bug comes in. Somehow before
our events were still in order, despite using a
dictionary for storing the eventful days in a
month with the dates as keys. Now after updating
and re-building after invalidation, the order is
wrong. To fix it, the collection structure is now
more JSON-like, with dicts wrapping arrays, not
replacing them, and dates and events / nested
dicts in separate arrays with matching indices.

This also greatly simplifies the Months screen's
dataSource, and the interface is slightly
modified. Also, the event manager no longer has a
public events setter.