Refine itemSize calculation in TileLayout.
Instead of hard-coding the number of columns
depending on orientation, which has a complexity
that grows with the number of different screen
sizes, use desiredItemSize, based on values in IB.
Column count becomes a direct result of trying to
fulfill desiredItemSize, and no longer requires
orientation.

Then move this whole thing into prepareForLayout,
so itemSize is correct in time for view changes
during orientation change, and manual calls to
updating the sizes become unneeded.

Minor reorganization included.