Fix crashes from missing TransitionInteractionDelegate.
Happens when dismissing to Months and Day screens.
Don't assume the delegate will be around; it has
to be weak, because the it ultimately creates a
strong reference to the transition. Just store a
strong reference to the window.

Setup and teardown of the interactive transition
are done during view controller appearance now, to
avoid having two pinch recognizers on the same
window.