Fix Months pinch gesture blocking Day pinch gesture
The solution to resolve the pinch recognizer
conflict is to look at the index for each
recognizer on their shared view/window's
recognizer list. The assumption is the parent
screen's recognizer is always added earlier.

Also allow simultaneous recognition for pinches to
allow both possible zoom in and out transitions.