Fix critical issue with wrong event durations.
All-day events were actually 2-day events, since we were unnecessarily
adjusting endDate when it auto-adjusted. This probably was a regression
from using the allDay flag. Leave it to the docs to not mention
anything.

Part-day events were not fully implemented, since their duration is
inferably 1 hour and not 24.

Also move the defaulting of such values out of validation and into
before-save.