Fix redundant shiftToInputView calls in EventVC.
This bug with NavigationTitleScrollView goes deep. updateVisibleItem
gets called upon scroll-view stop. But isSubviewVisible is wrong in if
contentOffset is where second item is visible, the first item also
counts as visible because of a shared boundary value during comparison.
The bug is hidden because updateVisibleItem sets for all candidates, and
the correct one is the last one. This also causes unwanted contentOffset
thrashing via visibleItem's didSet.

Finally, EventVC has a delegate method that calls
updateDayIdentifierToItem, which ultimately calls shiftToInputView on
the day picker.