Fix missing day label text when editing later event.
Setting minimum date in the navigationTitleScrollView:itemAtIndex:
delegate method that gets triggered during setUpDayMenu causes FormVC
KVO to be fired, so the later call to updateInputViewsWithFormDataObject
in viewDidLoad will get shorted from no value change, causing the change
handler that updates the label to not run.

To avoid deeply changing setUpDayMenu, the minimum date is now being set
in a newly renamed update method, and that method is being called in the
last part of state setup.

Minor refactoring included.