Fix regression with MonthHeaderView artifacts.
Was only able to verify the bug in the simulator. After scroll,
updateTitleView is called, but that seems to cause it to run during
rotation now and leave artifact views because we're dequeue-ing in the
method to get a relatively static measure.

Instead, just hard-code headerLabelTop and simplify. Also improve logging.