Rework CollectionViewTileCell tap animation.
Calling on a didSet is much riskier compared to using the appropriate
delegate methods. On device and unlike on simulator, the animations
stutter, regardless of technique. The original problem that yielded the
hack is solved by turning off UICollectionView.delaysContentTouches. But
note the stutter remains if the animation duration is bumped, so it's
likely there's a duration budget.

This also reworks the animation from a half-complete spring animation to
a two-part depress-reverse animation, without easing since duration is
short.

See http://stackoverflow.com/q/23392485.