Fixup 3a6c227, didSet not called during init...
...despite super being called before. Perhaps something to do with IUO.
This was causing bugs like the button to EKEventEditViewController not
showing up.

> The willSet and didSet observers of superclass properties are called
when a property is set in a subclass initializer, after the superclass
initializer has been called. They are not called while a class is
setting its own properties, before the superclass initializer has been
called.

Also: http://stackoverflow.com/q/25230780/65465